+++ GRDB/Core/Database.swift
@@ -292,7 +292,13 @@ public final class Database {
         }
     }
 
+    private var insideTransaction: Bool = false
+    
     private func beginTransaction(type: TransactionType = .Exclusive) throws {
+        guard insideTransaction == false else {
+            fatalError("Can not begin a transaction inside another transaction.")
+        }
+        
         switch type {
         case .Deferred:
             try execute("BEGIN DEFERRED TRANSACTION")
@@ -301,14 +307,26 @@ public final class Database {
         case .Exclusive:
             try execute("BEGIN EXCLUSIVE TRANSACTION")
         }
+        
+        insideTransaction = true
     }
     
     private func rollback() throws {
+        guard insideTransaction == true else {
+            fatalError("Can not rollback: no current transaction.")
+        }
+        
         try execute("ROLLBACK TRANSACTION")
+        insideTransaction = false
     }
     
     private func commit() throws {
+        guard insideTransaction == true else {
+            fatalError("Can not commit: no current transaction.")
+        }
+        
         try execute("COMMIT TRANSACTION")
+        insideTransaction = false
     }
 }